Исправлен баг с выводом карточек в favorites
